The sglr source is now under autoconfig.  A general cleanup has been performed, and the compiler now is invoked at a higher complaining level.

+++ sglr/sglr/parser.c
@@ -1,4 +1,4 @@
-/* 
+/*
   %% $Id$
 
   \section{Parser}
@@ -16,7 +16,7 @@
 
   \input{parser.htx}
 
-  \noindent\rule{3cm}{0.5pt} end parser.h 
+  \noindent\rule{3cm}{0.5pt} end parser.h
 
 */
 #include <TB.h>
@@ -27,7 +27,7 @@
 #include "stack.h"
 #include "tree-to-dot.h"
 
-/* 
+/*
    The following flags determine how much progress information is
    provided.
 */
@@ -37,14 +37,14 @@ extern FILE *log;
 extern int abbreviation_flag;
 extern int show_stack;
 extern int gc;
-/* 
+/*
    \subsection{Text Input}
 
    The input for the parser is a list of characters closed by an end
    of file symbol |EOF|. The list is accessed by means of a function
    that produces the next character in the list. The function is a
    parameter of the |next_token| function and can be determined by the
-   caller of the parser. 
+   caller of the parser.
 
    For the purpose of error messages we keep track of the current line
    and column.
@@ -53,11 +53,11 @@ int line;        /* current line */
 int col;         /* current column */
 int text_length; /* number of tokens read */
 #define eof 0
-/* 
+/*
    By definition we say that the start of the input is at
    column zero of line one.
 */
-void 
+void
 init_input(void)
 {
   line        = 1;
@@ -67,10 +67,10 @@ init_input(void)
 /*
   The function |next_token| reads the next token from the
   stream provided by the parameter function |get_next_char| and
-  updates the line and column values taking account of 
+  updates the line and column values taking account of
   newlines |\n|, tabs |\t| and end of file |EOF|.
 */
-int 
+int
 next_token(int(*get_next_char)(void))
 {
   int c;
@@ -85,7 +85,7 @@ next_token(int(*get_next_char)(void))
   if(debugflag) fprintf(log, "next_token = %c\n", c);
   return c;
 }
-/* 
+/*
    \subsection{Shift Pairs}
 
    A shift pair is a pair of a state and a stack and is used to
@@ -101,7 +101,7 @@ typedef term shift_pair;
 
 #define add_shift_pair(st, a, next) next = new_shift_pair(st, a, next)
 
-/* 
+/*
    \subsection{Paths}
 
    A path is a pair of a stack and a list of terms. A path represents
@@ -147,19 +147,19 @@ find_paths(stack *st, int i, st_link *l0, bool link_seen, term *sons)
   term *paths = NULL;
   term *newsons = NULL;
 
-  if (debugflag)  fprintf(log, "find_paths(%d, %d, x, %d, %d)\n", 
-			  STATE(st), i, link_seen, sons);
-  if (st == NULL) 
+  if (debugflag)  fprintf(log, "find_paths(%d, %d, x, %d, %d)\n",
+			  STATE(st), i, link_seen, (int) sons);
+  if (st == NULL)
     paths = NULL;
-  else if (i == 0 && link_seen) 
+  else if (i == 0 && link_seen)
     paths = new_path(st, sons, NULL);
   else if (i > 0)
-    for (ls = LINKS(st); ls != NULL; ls = next(ls)) 
+    for (ls = LINKS(st); ls != NULL; ls = next(ls))
       {
 	l1 = first(ls);
 	newsons = cons(TREE(l1), sons);
-	paths = conc_paths(find_paths(STACK(l1), i - 1, l0, 
-				      link_seen || (l0 == l1), 
+	paths = conc_paths(find_paths(STACK(l1), i - 1, l0,
+				      link_seen || (l0 == l1),
 				      newsons),
 			   paths);
       }
@@ -169,16 +169,16 @@ find_paths(stack *st, int i, st_link *l0, bool link_seen, term *sons)
    \subsection{The Interpreter}
 
    The state of the parser is stored in the following global variables.
-   
+
 */
 parse_table *table;
 int          current_token;
-stack       *accepting_stack;  
-stacks      *active_stacks;    
-stacks      *for_actor;         
-stacks      *for_actor_delayed;         
-shift_pair  *for_shifter;       
-/* 
+stack       *accepting_stack;
+stacks      *active_stacks;
+stacks      *for_actor;
+stacks      *for_actor_delayed;
+shift_pair  *for_shifter;
+/*
    Predeclaration of the auxiliary functions.
 */
 void parse_character(void);
@@ -188,7 +188,7 @@ void reducer(stack *, state, label, term_list *, bool);
 void do_limited_reductions(stack*, action*, st_link*);
 void shifter(void);
 term *result(void);
-/* 
+/*
    \paragraph{Parse}
 
    The parse function |parse| parses a text with a parse table |ptable|.
@@ -208,7 +208,7 @@ term *result(void);
 term *
 parse(parse_table *ptable, int(*get_next_char)(void))
 {
-  init_input(); 
+  init_input();
   table = ptable;
 
   accepting_stack = NULL;
@@ -216,7 +216,7 @@ parse(parse_table *ptable, int(*get_next_char)(void))
   TBprotect(&active_stacks);
   TBprotect(&accepting_stack);
 
-  do 
+  do
     {
       if(show_stack) stack_to_dotfile(active_stacks);
       current_token = next_token(get_next_char);
@@ -226,11 +226,11 @@ parse(parse_table *ptable, int(*get_next_char)(void))
       if(show_stack) stack_to_dotfile_end();
     }
   while (current_token != eof && active_stacks != NULL);
-    
+
   if(show_stack) stack_to_dotfile(mk_list(accepting_stack, NULL));
   if(show_stack) stack_to_dotfile_end();
   return result();
-} 
+}
 /*
   \paragraph{Termination}
 
@@ -240,34 +240,34 @@ parse(parse_table *ptable, int(*get_next_char)(void))
   direct link to the initial state. The tree on this link is the parse
   forest with all possible parse trees.  If parsing failed an error
   term is returned. A distinction is made between an error at end of
-  file an error in the middle of the file. 
+  file an error in the middle of the file.
 */
 term *
 result(void)
 {
-  if (accepting_stack != NULL) 
+  if (accepting_stack != NULL)
     {
       term *t;
       t = TREE(first(LINKS(accepting_stack)));
       return TBmake("parsetree(%t,%d)", t, nr_ambiguities);
-    } 
-  else  
+    }
+  else
     return TBmake("parse-error([character(%d), line(%d),"
 		  "col(%d), char(%d)])",
 		  current_token, line, col, text_length);
-    
+
 }
-/* 
+/*
    \paragraph{Parse Character}
 
    For each stack on |active_stacks| handle the actions for the
-   current token. 
+   current token.
 */
 
-void 
+void
 parse_character(void)
 {
-  stack *st;      
+  stack *st;
 
   for_actor = active_stacks;
   for_actor_delayed = NULL;
@@ -279,12 +279,12 @@ parse_character(void)
     else
       pop(st, for_actor);
     if(!rejected(st))
-      actor(st); 
+      actor(st);
     if(show_stack) links_to_dot(stack_dot, st);
   }
 }
 
-/* 
+/*
    \paragraph{Actor}
 
    Handle the actions for stack |st| and the current token.  A reduce
@@ -299,8 +299,8 @@ parse_character(void)
 */
 void
 actor(stack *st)
-{ 
-  actions *as; 
+{
+  actions *as;
   action *a;
   as = ACTIONS(table, STATE(st), current_token);
   while(pop(a, as))
@@ -313,7 +313,7 @@ actor(stack *st)
       break;
     case ERROR:
       break;
-    } 
+    }
 }
 /*
   \paragraph{Reductions}
@@ -324,17 +324,17 @@ actor(stack *st)
   and the trees along the path as arguments. The new tree is the link
   from a new stack to the stack at the end of the path.
 */
-void 
+void
 do_reductions(stack *st, action *a)
 {
-  path  *p; 
+  path  *p;
   paths *ps;
-  label prod; 
+  label prod;
   prod = A_PROD(a);
 
   ps = find_paths(st, A_NR_ARGS(a), NULL, TRUE, NULL);
-  while(pop(p, ps)) 
-    reducer(P_STACK(p), GOTO(table, STATE(P_STACK(p)), prod), 
+  while(pop(p, ps))
+    reducer(P_STACK(p), GOTO(table, STATE(P_STACK(p)), prod),
 	    prod, P_ARGS(p), A_REJECT(a));
 }
 
@@ -342,8 +342,8 @@ do_reductions(stack *st, action *a)
   \paragraph{Reducer}
 
   For each path in |p| construct the parse tree with the list of
-  descendants found and create a new stack. 
-  
+  descendants found and create a new stack.
+
   Stack |st0| is the stack at the end of path |p|. State |s| is the
   next state when reducing production |r| from the state of
   |st0|. Look for a stack |st1| on |active_stacks| with state |s|.  If
@@ -368,49 +368,49 @@ do_reductions(stack *st, action *a)
 void
 reducer(stack *st0, state s, label prod, term_list *kids, bool reject)
 {
-  term* t; 
+  term* t;
   st_link *nl;
-  stack *st1; 
+  stack *st1;
 
   t = apply(table, prod, kids);
 
-  if(debugflag) 
+  if(debugflag)
     TBprintf(log, "reducing %t\n", dot_trm_yield(t));
 
   /* new stack */
-  if((st1 = find_stack(s, active_stacks)) == NULL) { 
-    st1 = new_stack(s, NULL); 
+  if((st1 = find_stack(s, active_stacks)) == NULL) {
+    st1 = new_stack(s, NULL);
     add_link(st1, t, st0, nl);
     add_stack(st1, active_stacks);
     add_stack(st1, for_actor_delayed);
-    if (reject) mark_stack_rejected(st1, nl); 
+    if (reject) mark_stack_rejected(st1, nl);
   }
 
   /* ambiguity */
   else if((nl = find_direct_link(st1, st0)) != NULL) {
-    if(debugflag) 
+    if(debugflag)
       TBprintf(log, "direct link found\n");
     amb(TREE(nl), t);
-    if (reject) mark_link_rejected1(st1, nl);	  
+    if (reject) mark_link_rejected1(st1, nl);
   }
 
   /* add new direct link from |st1| to |st0| and
      recheck all reductions all for |st1|. */
-  else { 
-    stacks *sts = NULL; 
-    stack  *st2 = NULL; 
-    add_link(st1, t, st0, nl);  
+  else {
+    stacks *sts = NULL;
+    stack  *st2 = NULL;
+    add_link(st1, t, st0, nl);
     if (reject) mark_link_rejected2(st1, nl);
     sts = active_stacks;
-    while(pop(st2, sts)) { 
+    while(pop(st2, sts)) {
       actions *as;
       action *a;
       if(!rejected(st2) && !in_stacks(st2, for_actor)
 	 && !in_stacks(st2, for_actor_delayed)) {
 	as = ACTIONS(table, STATE(st2), current_token);
 	while(pop(a, as))
-	  if(A_KIND(a) == REDUCE) 
-	    do_limited_reductions(st2, a, nl);  
+	  if(A_KIND(a) == REDUCE)
+	    do_limited_reductions(st2, a, nl);
       }
     }
   }
@@ -420,20 +420,20 @@ reducer(stack *st0, state s, label prod, term_list *kids, bool reject)
 
   Only those reductions that have a path containing link |l|
 */
-void 
+void
 do_limited_reductions(stack *st, action *a, st_link *l)
 {
-  path  *p; 
+  path  *p;
   paths *ps;
-  label prod; 
+  label prod;
   prod = A_PROD(a);
 
   ps = find_paths(st, A_NR_ARGS(a), l, FALSE, NULL);
-  while(pop(p, ps)) 
-    reducer(P_STACK(p), GOTO(table, STATE(P_STACK(p)), prod), 
+  while(pop(p, ps))
+    reducer(P_STACK(p), GOTO(table, STATE(P_STACK(p)), prod),
 	    prod, P_ARGS(p), A_REJECT(a));
 }
-/* 
+/*
    \paragraph{Shifter}
 
    Do all shifts in |for_shifter|. For each shift pair |(st1, s)|, a
@@ -445,15 +445,15 @@ do_limited_reductions(stack *st, action *a, st_link *l)
 void
 shifter(void)
 {
-  term *t;   
+  term *t;
   stack *st0, *st1;
   term *shift_pair;
   state s;
   st_link *l;
-  
+
   active_stacks = NULL;
   t = PROD(table, current_token);
-  
+
   while(pop(shift_pair, for_shifter)) {
     s = SP_STATE(shift_pair);
     st0 = SP_STACK(shift_pair);
@@ -461,7 +461,7 @@ shifter(void)
       st1 = new_stack(s, NULL);
       add_stack(st1, active_stacks);
     }
-    add_link(st1, t, st0, l);	
+    add_link(st1, t, st0, l);
     if(show_stack) links_to_dot(stack_dot, st1);
   }
 } /* shifter */
